@@ -1,8 +1,7 @@
 import EmployeesListItem from "../employeesListItem/EmployeesListItem";
-
 import './employeesList.css';
 
-const EmployeesList = ({ data, onDelete, onToggleProp }) => {
+const EmployeesList = ({ data, onDeleteItem, onToggleProp, onChangeSalary }) => {
     
     const elements = data.map(item => {
         const { id, ...itemProps } = item;
@@ -10,7 +9,8 @@ const EmployeesList = ({ data, onDelete, onToggleProp }) => {
             <EmployeesListItem 
                 key={id} 
                 {...itemProps} 
-                onDelete={() => onDelete(id)} 
+                onDeleteItem={() => onDeleteItem(id)} 
+                onChangeSalary={(event) => onChangeSalary(id, event.target.value)}
                 onToggleProp={(event) => onToggleProp(id, event.currentTarget.getAttribute('data-toggle'))} />
         );
     });
